feat: decode appearance update block
The appearance block is really big - probably the size of everything
else in this PR combined. So rather than creating messages/decoders for
it now, I've chosen to just "read all remaining bytes on the buffer" and
pass them off to the original implementation.

We can refactor this to have a proper structure in a future PR.